changed particle sim page to main streamlit app file
